pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "relayroute"
version = "1.0.0"
description = "Zone-based relay delivery infrastructure API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "scikit-learn>=1.4.0",
    "networkx>=3.2",
    "scipy>=1.12.0",
    "openai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["relayroute*"]
